backend-api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.3
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.7.0
et_xmlfile==2.0.0
fastapi==0.115.12
Flask==3.1.1
fonttools==4.58.1
fpdf==1.7.2
fpdf2==2.8.3
greenlet==3.2.3
h11==0.16.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.41.1
numpy==2.2.6
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.0
pdf2image==1.17.0
pillow==11.2.1
playwright==1.52.0
plotly==6.1.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydyf==0.11.0
pyee==13.0.0
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
pyparsing==3.2.3
pypdf==5.6.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
ssh-import-id==5.11
starlette==0.46.2
supervisor==4.2.1
svglib==1.5.1
tabulate==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
uvicorn==0.34.3
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
